torii.lib: Moved `torii.lib.soc.memory` into `torii.lib.mem.map`
This moves the memory map constructs into the new `torii.lib.mem` module so when we pull the bus definitions out we can not need to depend on SoC specific things where we don't want them.

This also adds the deprecation stubs for `torii.lib.soc.memory`.
