+++ docs/library/mem/map.md
@@ -1,0 +1,8 @@
+# Memory Map
+
+```{eval-rst}
+
+.. autoclass:: torii.lib.mem.map.ResourceInfo
+.. autoclass:: torii.lib.mem.map.MemoryMap
+
+```
