+++ docs/library/soc/index.md
@@ -6,7 +6,6 @@ The {py:mod}`torii.lib.soc` module provides various modules for use in SoCs and 
 :maxdepth: 2
 
 event
-memory
 periph
 
 csr/index
