--- docs/library/soc/memory.md
@@ -1,8 +1,0 @@
-# Memory Map
-
-```{eval-rst}
-
-.. autoclass:: torii.lib.soc.memory.ResourceInfo
-.. autoclass:: torii.lib.soc.memory.MemoryMap
-
-```
