+++ tests/lib/mem/test_map.py
@@ -1,8 +1,8 @@
 # SPDX-License-Identifier: BSD-2-Clause
 
-from torii.lib.soc.memory import MemoryMap, ResourceInfo, _RangeMap
+from torii.lib.mem.map import MemoryMap, ResourceInfo, _RangeMap
 
-from ...utils             import ToriiTestSuiteCase
+from ...utils          import ToriiTestSuiteCase
 
 class RangeMapTestCase(ToriiTestSuiteCase):
 	def test_insert(self):
@@ -354,7 +354,7 @@ class MemoryMapTestCase(ToriiTestSuiteCase):
 		with self.assertRaisesRegex(
 			ValueError,
 			r'Memory map has been frozen. Cannot add window '
-			r'<torii\.lib\.soc\.memory\.MemoryMap object at .+?>'
+			r'<torii\.lib\.mem\.map\.MemoryMap object at .+?>'
 		):
 			memory_map.add_window(MemoryMap(addr_width = 1, data_width = 8))
 
@@ -408,7 +408,7 @@ class MemoryMapTestCase(ToriiTestSuiteCase):
 		with self.assertRaisesRegex(
 			ValueError,
 			r'Address range 0x200\.\.0x600 overlaps with window '
-			r'<torii\.lib\.soc\.memory\.MemoryMap object at .+?> at 0x0\.\.0x400'
+			r'<torii\.lib\.mem\.map\.MemoryMap object at .+?> at 0x0\.\.0x400'
 		):
 			memory_map.add_window(MemoryMap(addr_width = 10, data_width = 8), addr = 0x200)
 
@@ -418,7 +418,7 @@ class MemoryMapTestCase(ToriiTestSuiteCase):
 		memory_map.add_window(window)
 		with self.assertRaisesRegex(
 			ValueError,
-			r'Window <torii\.lib\.soc\.memory\.MemoryMap object at .+?> is already added '
+			r'Window <torii\.lib\.mem\.map\.MemoryMap object at .+?> is already added '
 			r'at address range 0x0\.\.0x400'
 		):
 			memory_map.add_window(window)
