+++ tests/lib/soc/csr/test_bus.py
@@ -5,7 +5,7 @@ from torii.hdl.dsl         import Module
 from torii.hdl.rec         import Layout
 from torii.hdl.ir          import Fragment
 from torii.lib.soc.csr.bus import Decoder, Element, Interface, Multiplexer
-from torii.lib.soc.memory  import MemoryMap
+from torii.lib.mem.map     import MemoryMap
 from torii.sim             import Delay, Simulator
 
 from ....utils             import ToriiTestSuiteCase
