+++ torii/lib/soc/csr/bus.py
@@ -12,7 +12,7 @@ from ....hdl.dsl     import Module
 from ....hdl.ir      import Elaboratable
 from ....hdl.rec     import Record
 from ....util.units  import log2_ceil
-from ..memory        import MemoryMap
+from ...mem.map      import MemoryMap
 
 __all__ = (
 	'Decoder',
